# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trace - AI-powered collaborative memory API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectStatsMemoriesByType } from './projectStatsMemoriesByType';

export interface ProjectStats {
  totalMemories: number;
  totalTasks: number;
  openTasks: number;
  memoriesByType: ProjectStatsMemoriesByType;
  recentDecisions?: number;
  completedTasks?: number;
}