# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trace - AI-powered collaborative memory API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputStatus } from './taskInputStatus';

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  assignee?: string;
  status?: TaskInputStatus;
  dueDate?: Date;
  memoryId?: number;
}